gthread: Remove locks + general clean-up
This change-set aims change so that we do everything non-related to
the request on the main thread. This way, we can hopefully avoid lots
of strange race conditions and make it easier to reason about the code.

This also fixes #3146, but there are leaner ways of doing that.

Work in progress.